removed html table. Now the only grid available is css-grid
