--- src/components/format-grid/styles.scss.d.ts
@@ -1,7 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-
-}
-declare var cssExports: CssExports;
-export = cssExports;
