+++ src/components/uxl-grid/template.ts
@@ -1,17 +1,81 @@
 import { html } from "lit-element/lit-element";
 import { UxlGrid} from "./uxl-grid";
 import {iconTemplate} from "../../icons/icons";
-import "../format-table/format-table";
-import "../format-grid/format-grid";
+import {repeat} from "lit-html/directives/repeat";
+import {nothing} from "lit-html";
+import {classMap} from "lit-html/directives/class-map";
 
 
 export const template = (props: UxlGrid) => html`
   ${iconTemplate()}
-  ${props.isFormatGrid ?
-        html`
-            <format-grid .source="${props.source}" .columns="${props.columns}" id="grid" part="format-grid" .showHeader="${props.showHeader}" .numberColumns="${props.columns.length}"
-            exportparts="header, header__cell-1, header__cell-2, header__cell-3, header__cell-4, header__cell-5, header__cell-6, header__cell-7, content, content__row, content__cell, content__cell-1, content__cell-2, content__cell-3, content__cell-4, content__cell-5, content__cell-6, content__cell-7"></format-grid>` :
-		html `
-		    <format-table .source="${props.source}" .columns="${props.columns}" .showHeader="${props.showHeader}"></format-table>
-		`}
-`;
+    <div id="grid">
+      ${props.showHeader ? html`
+	<div class="header" part="header">
+	${repeat(
+    props.columns,
+    (column, index) => html`
+      <div
+        id="header-cell-${index + 1}"
+        class="header__cell"
+        data-column="${JSON.stringify(column)}"
+        part="header__cell-${index + 1}"
+      >
+        ${column.displayName || ""}
+        ${
+        column.order == "ASC"
+            ? html`
+                <iron-icon class="order" icon="uxl-grid:arrow-up"></iron-icon>
+              `
+            : html`
+                ${
+                column.order == "DES"
+                    ? html`
+                        <iron-icon
+                          class="order"
+                          icon="uxl-grid:arrow-down"
+                        ></iron-icon>
+                      `
+                    : nothing
+            }
+              `
+    }
+      </div>
+    `
+)}
+	</div>
+` : nothing}
+<div class="content" part="content" id="content">
+${repeat(
+    props.orderedList,
+    (item, index) => html`
+          <div
+            id="row-${index + 1}"
+            class="content__row ${classMap({disabled: item.disabled})}"
+            part="content__row"
+          >
+            ${
+        repeat(
+            props.columns,
+            (column, index) => html`
+                  <div
+                    id="column-${index + 1}"
+                    class="content__cell"
+                    part="content__cell, content__cell-${index + 1}"
+                    data-item="${JSON.stringify(item)}"
+                  >
+                    ${item[column.property] || ""}
+                  </div>
+                `
+        )
+    }
+          </div>
+        `
+)}
+  </div>
+</div>
+</div>
+  
+`
+
+
+
